Defining Book as a third-party block
Both as exercise and to avoid need for block tests, which would distract
from the object of this PR (meta support).
