+++ blocks/library/index.js
@@ -22,4 +22,3 @@ import './text-columns';
 import './verse';
 import './video';
 import './audio';
-import './book';
