+++ gutenberg.php
@@ -29,4 +29,7 @@ if ( gutenberg_can_init() ) {
 	foreach ( glob( dirname( __FILE__ ) . '/blocks/library/*/index.php' ) as $block_logic ) {
 		require_once $block_logic;
 	}
+
+	// FIXME don't merge this
+	require_once dirname( __FILE__ ) . '/hypothetical-book-block.php';
 }
